feat: em.clone enhancements
